requirements.txt
jsonschema==3.0.1
PyYAML==5.1.2
